settings: refactored folder selection models and templates
