+++ ClrVpin/Controls/Folder/ContentFolderType.xaml
@@ -1,0 +1,53 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+                    xmlns:validationRules="clr-namespace:ClrVpin.Validation_Rules"
+                    xmlns:b="http://schemas.microsoft.com/xaml/behaviors"
+                    xmlns:controls="clr-namespace:ClrVpin.Controls"
+                    xmlns:folder="clr-namespace:ClrVpin.Controls.Folder">
+
+  <ResourceDictionary.MergedDictionaries>
+    <controls:SharedResourceDictionary Source="pack://application:,,,/MaterialDesignThemes.Wpf;component/Themes/MaterialDesignTheme.Defaults.xaml" />
+    <controls:SharedResourceDictionary Source="\Controls\Styles.xaml" />
+    <controls:SharedResourceDictionary Source="FolderSelection.xaml" />
+  </ResourceDictionary.MergedDictionaries>
+
+  <Style x:Key="TextBoxStyle" TargetType="TextBox" BasedOn="{StaticResource MaterialDesignFloatingHintTextBox}">
+    <Setter Property="Width" Value="400" />
+    <Setter Property="materialDesign:TextFieldAssist.HasClearButton" Value="True" />
+  </Style>
+
+  <!-- present folder selection with file extension and kindred file extension support -->
+  <DataTemplate x:Key="ContentFolderTypeModelTemplate" DataType="folder:ContentFolderTypeModel">
+    <WrapPanel Margin="0,0,0,0" Height="50">
+      <ContentControl ContentTemplate="{StaticResource FolderSelection}" Content="{Binding}" />
+
+      <TextBox Margin="5,0,0,0" Width="150" Style="{StaticResource TextBoxStyle}" VerticalAlignment="Top" materialDesign:HintAssist.Hint="File(s)"
+               ToolTip="Only files with matching extensions will be considered. Extensions are ordered in descending priority order.">
+        <TextBox.Text>
+          <Binding Path="Extensions" UpdateSourceTrigger="PropertyChanged">
+            <Binding.ValidationRules>
+              <validationRules:NotEmptyValidationRule ValidatesOnTargetUpdated="True" />
+            </Binding.ValidationRules>
+          </Binding>
+        </TextBox.Text>
+        <b:Interaction.Triggers>
+          <b:EventTrigger EventName="TextChanged">
+            <b:InvokeCommandAction Command="{Binding ExtensionsChangedCommandWithParam}" PassEventArgsToCommand="True" />
+          </b:EventTrigger>
+        </b:Interaction.Triggers>
+      </TextBox>
+
+      <TextBox Margin="5,0,0,0" Width="150" Style="{StaticResource TextBoxStyle}" VerticalAlignment="Top" materialDesign:HintAssist.Hint="Linked Files(s)"
+               ToolTip="Any related files with matching extensions will also be moved/copied/deleted."
+               Text="{Binding KindredExtensions, UpdateSourceTrigger=PropertyChanged}">
+        <b:Interaction.Triggers>
+          <b:EventTrigger EventName="TextChanged">
+            <b:InvokeCommandAction Command="{Binding KindredExtensionsChangedCommandWithParam}" PassEventArgsToCommand="True" />
+          </b:EventTrigger>
+        </b:Interaction.Triggers>
+      </TextBox>
+    </WrapPanel>
+  </DataTemplate>
+
+</ResourceDictionary>
