+++ ClrVpin/Settings/Settings.xaml
@@ -1,15 +1,14 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:settings="clr-namespace:ClrVpin.Settings"
-                    xmlns:b="http://schemas.microsoft.com/xaml/behaviors"
                     xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-                    xmlns:validationRules="clr-namespace:ClrVpin.Validation_Rules"
                     xmlns:controls="clr-namespace:ClrVpin.Controls">
 
   <ResourceDictionary.MergedDictionaries>
     <controls:SharedResourceDictionary Source="\Controls\Converters.xaml" />
     <controls:SharedResourceDictionary Source="\Controls\Styles.xaml" />
-    <controls:SharedResourceDictionary Source="\Controls\FolderSelection\FolderSelection.xaml" />
+    <controls:SharedResourceDictionary Source="\Controls\Folder\FolderSelection.xaml" />
+    <controls:SharedResourceDictionary Source="\Controls\Folder\ContentFolderType.xaml" />
   </ResourceDictionary.MergedDictionaries>
 
   <!-- items control with virtualisation and scroll viewer  - https://stackoverflow.com/questions/2783845/virtualizing-an-itemscontrol-->
@@ -42,38 +41,6 @@
     </Setter>
   </Style>
 
-  <DataTemplate x:Key="ContentTypeModelTemplate" DataType="settings:ContentTypeModel">
-    <WrapPanel Margin="0,0,0,0" Height="50">
-      <ContentControl ContentTemplate="{StaticResource FolderSelection}" Content="{Binding}" />
-
-      <TextBox Margin="5,0,0,0" Width="150" Style="{StaticResource TextBoxStyle}" VerticalAlignment="Top" materialDesign:HintAssist.Hint="File(s)"
-               ToolTip="Only files with matching extensions will be considered. Extensions are ordered in descending priority order.">
-        <TextBox.Text>
-          <Binding Path="Extensions" UpdateSourceTrigger="PropertyChanged">
-            <Binding.ValidationRules>
-              <validationRules:NotEmptyValidationRule ValidatesOnTargetUpdated="True" />
-            </Binding.ValidationRules>
-          </Binding>
-        </TextBox.Text>
-        <b:Interaction.Triggers>
-          <b:EventTrigger EventName="TextChanged">
-            <b:InvokeCommandAction Command="{Binding ExtensionsChangedCommandWithParam}" PassEventArgsToCommand="True" />
-          </b:EventTrigger>
-        </b:Interaction.Triggers>
-      </TextBox>
-
-      <TextBox Margin="5,0,0,0" Width="150" Style="{StaticResource TextBoxStyle}" VerticalAlignment="Top" materialDesign:HintAssist.Hint="Linked Files(s)"
-               ToolTip="Any related files with matching extensions will also be moved/copied/deleted."
-               Text="{Binding KindredExtensions, UpdateSourceTrigger=PropertyChanged}">
-        <b:Interaction.Triggers>
-          <b:EventTrigger EventName="TextChanged">
-            <b:InvokeCommandAction Command="{Binding KindredExtensionsChangedCommandWithParam}" PassEventArgsToCommand="True" />
-          </b:EventTrigger>
-        </b:Interaction.Triggers>
-      </TextBox>
-    </WrapPanel>
-  </DataTemplate>
-
   <DataTemplate x:Key="SettingsTemplate" DataType="{x:Type settings:SettingsViewModel}">
     <ScrollViewer VerticalScrollBarVisibility="Auto">
       <StackPanel Margin="5">
@@ -89,7 +56,7 @@
                 <materialDesign:PackIcon Kind="MagicWand" />
               </Button>
             </WrapPanel>
-            <ItemsControl ItemsSource="{Binding PinballContentTypeModels}" BorderThickness="0" ItemTemplate="{StaticResource ContentTypeModelTemplate}" />
+            <ItemsControl ItemsSource="{Binding PinballContentTypeModels}" BorderThickness="0" ItemTemplate="{StaticResource ContentFolderTypeModelTemplate}" />
           </StackPanel>
         </GroupBox>
 
@@ -105,7 +72,7 @@
                 <materialDesign:PackIcon Kind="MagicWand" />
               </Button>
             </WrapPanel>
-            <ItemsControl ItemsSource="{Binding FrontendContentTypeModels}" BorderThickness="0" ItemTemplate="{StaticResource ContentTypeModelTemplate}"
+            <ItemsControl ItemsSource="{Binding FrontendContentTypeModels}" BorderThickness="0" ItemTemplate="{StaticResource ContentFolderTypeModelTemplate}"
                           Style="{StaticResource ItemsControlVirtualizedStyle}" MaxHeight="375" />
           </StackPanel>
         </GroupBox>
